chore: checkpoint project state
Includes build-fix work, root documentation refresh, doc gate outputs, and concurrent workspace changes.

+++ Assets/_Project/Scripts/Physics/Buoyancy/AsyncReadback/AsyncBuoyancyReadbackJobs.cs
@@ -187,7 +187,7 @@ namespace Hecton8.Physics
             }
 
             double heightAupY = CameraAupY + localHeight;
-            bool finite = math.isfinite(localHeight) && math.isfinite((float)heightAupY);
+            bool finite = math.isfinite(localHeight) && math.isfinite(heightAupY);
             if (!finite)
             {
                 localHeight = 0f;
